requirements.txt
numpy==1.24.2
openai==0.27.2
pandas==1.5.3
python-dotenv==1.0.0
python-telegram-bot[callback-data]
python-telegram-bot[job-queue]
PyYAML==6.0
sentence_transformers==2.2.2
Telethon==1.27.0
tiktoken==0.3.3
torch==2.0.0
weaviate_client==3.15.4
